Move staging ingestion tests to a mixin and skip them
